PSY-933: lazy-load the admin nav off the public bundle
Per the Lighthouse-gate review: the admin nav (adminNav config + the 7
queue-count hooks) was statically imported by the global Sidebar + TopBar, so
every page — including anonymous /explore — downloaded admin-only code. Split it
into dynamically-imported chunks loaded only when an admin is on the /admin
shell:

- SidebarNavLink: extracted the shared link row (active / badge / collapse /
  tooltip) so the public Sidebar and the admin rail render identically from one
  source.
- AdminSidebarNav (default export, next/dynamic ssr:false): the desktop admin
  rail — adminNav config + useAdminNavCounts({enabled:true}) live here, so the
  module is a /admin-only chunk. Mounted only when isAdmin && pathname==='/admin'.
- AdminDrawerNav (default export, dynamic): the mobile-drawer equivalent.
- Sidebar / TopBar no longer statically import adminNav, useAdminNavCounts, or
  useSearchParams — those move into the lazy chunks. They keep the same
  showAdminNav gate and now render <AdminSidebarNav> / <AdminDrawerNav>.

Behavior is unchanged (the count queries are still gated — now by mount, which
is strictly stronger: non-admins/public routes never even load the module).
Tests restructured: the admin-rail/drawer behavior moves to AdminSidebarNav.test
+ AdminDrawerNav.test (rendered directly); Sidebar.test/TopBar.test keep their
public-nav suites + a delegation test (next/dynamic stubbed). Full suite green
(5284). Note: this is good hygiene + removes this PR's /explore footprint; the
gate's residual flake is environmental (cold-preview TTI median > 2.5s budget).
