+++ frontend/components/layout/Sidebar.tsx
@@ -1,11 +1,11 @@
 'use client'
 
-import Link from 'next/link'
-import { usePathname, useSearchParams } from 'next/navigation'
+import { usePathname } from 'next/navigation'
+import dynamic from 'next/dynamic'
 import {
   Calendar, Mic2, MapPin, Disc3, Tag, Tags, Tent, BookOpen, Headphones, Newspaper,
   Send, Library, LayoutList, MessageSquarePlus, UserCircle, Shield, PanelLeftClose, PanelLeft,
-  ExternalLink, Globe, TrendingUp, Bell, HeartHandshake, Trophy, Radio, Music, Compass, ArrowLeft,
+  Globe, TrendingUp, Bell, HeartHandshake, Trophy, Radio, Music, Compass,
 } from 'lucide-react'
 import type { LucideIcon } from 'lucide-react'
 import { cn } from '@/lib/utils'
@@ -13,10 +13,11 @@ import { useAuthContext } from '@/lib/context/AuthContext'
 import {
   Tooltip, TooltipContent, TooltipProvider, TooltipTrigger,
 } from '@/components/ui/tooltip'
-import {
-  adminNavGroups, adminTabHref, isAdminTabActive, ADMIN_BADGE_CLASS,
-} from './adminNav'
-import { useAdminNavCounts } from '@/lib/hooks/admin/useAdminNavCounts'
+import { SidebarNavLink } from './SidebarNavLink'
+
+// The admin rail (config + the 7 queue-count hooks) is a separate chunk loaded
+// only when an admin is on the /admin shell — public pages never download it.
+const AdminSidebarNav = dynamic(() => import('./AdminSidebarNav'), { ssr: false })
 
 export interface SidebarNavItem {
   href: string
@@ -70,94 +71,35 @@ interface SidebarProps {
 
 export function Sidebar({ collapsed, onToggleCollapse }: SidebarProps) {
   const pathname = usePathname()
-  const searchParams = useSearchParams()
   const { user, isAuthenticated } = useAuthContext()
 
   // In the admin area the rail becomes context-aware (PSY-933): it swaps the
-  // public Discover/Community groups for the grouped admin sections. Gated on
-  // isAdmin so a non-admin mid-redirect at /admin still sees the public nav.
-  // Scoped to the exact /admin tab-shell (usePathname() strips the ?tab= query,
-  // so this still matches /admin?tab=…). Standalone /admin/<section> sub-routes
-  // keep the public nav — their pre-PSY-933 behavior — rather than mis-rendering
-  // a ?tab=-only active state; a path-aware rail for those is a follow-up.
+  // public Discover/Community groups for the grouped admin sections (rendered by
+  // the lazily-loaded AdminSidebarNav). Gated on isAdmin so a non-admin
+  // mid-redirect at /admin still sees the public nav, and scoped to the exact
+  // /admin tab-shell (usePathname() strips the ?tab= query, so this still matches
+  // /admin?tab=…). Standalone /admin/<section> sub-routes keep the public nav —
+  // their pre-PSY-933 behavior — rather than mis-rendering a ?tab=-only active
+  // state; a path-aware rail for those is a follow-up.
   const isAdmin = !!user?.is_admin
   const showAdminNav = isAdmin && pathname === '/admin'
-  const tabParam = searchParams.get('tab')
-  // Gated by showAdminNav: these admin-only count queries must not fire on
-  // public routes or for non-admins (they'd 403 / waste requests).
-  const counts = useAdminNavCounts({ enabled: showAdminNav })
 
   const isActive = (href: string) => {
     if (href === '/') return pathname === '/'
     return pathname === href || pathname.startsWith(href + '/')
   }
 
-  // Unified link row used by both the public and admin nav. `badge`, when set,
-  // renders a count pill (expanded) or a corner dot (collapsed).
-  const renderNavLink = ({
-    keyVal, href, label, icon: Icon, active, external = false,
-    badge = null,
-  }: {
-    keyVal: string
-    href: string
-    label: string
-    icon: LucideIcon
-    active: boolean
-    external?: boolean
-    badge?: { count: number; className: string } | null
-  }) => {
-    const showBadge = badge != null && badge.count > 0
-    const link = (
-      <Link
-        href={href}
-        target={external ? '_blank' : undefined}
-        rel={external ? 'noopener noreferrer' : undefined}
-        className={cn(
-          'flex items-center gap-3 rounded-md px-3 py-2 text-sm font-medium transition-colors',
-          active
-            ? 'bg-sidebar-accent text-sidebar-accent-foreground'
-            : 'text-sidebar-foreground/70 hover:bg-sidebar-accent/50 hover:text-sidebar-accent-foreground',
-          collapsed && 'relative justify-center px-2'
-        )}
-      >
-        <Icon className="h-4 w-4 shrink-0" />
-        {!collapsed && <span>{label}</span>}
-        {!collapsed && external && (
-          <ExternalLink className="ml-auto h-3 w-3 opacity-50" />
-        )}
-        {!collapsed && showBadge && (
-          <span className={cn('ml-auto rounded-full px-2 py-0.5 text-xs font-medium text-white', badge!.className)}>
-            {badge!.count}
-          </span>
-        )}
-        {collapsed && showBadge && (
-          <span className={cn('absolute right-1.5 top-1.5 h-2 w-2 rounded-full', badge!.className)} aria-hidden />
-        )}
-      </Link>
-    )
-
-    if (collapsed) {
-      return (
-        <Tooltip key={keyVal}>
-          <TooltipTrigger asChild>{link}</TooltipTrigger>
-          <TooltipContent side="right">
-            {label}{showBadge ? ` (${badge!.count})` : ''}
-          </TooltipContent>
-        </Tooltip>
-      )
-    }
-    return <div key={keyVal}>{link}</div>
-  }
-
-  const renderItem = (item: SidebarNavItem) =>
-    renderNavLink({
-      keyVal: item.href,
-      href: item.href,
-      label: item.label,
-      icon: item.icon,
-      active: !item.external && isActive(item.href),
-      external: item.external,
-    })
+  const renderItem = (item: SidebarNavItem) => (
+    <SidebarNavLink
+      key={item.href}
+      href={item.href}
+      label={item.label}
+      icon={item.icon}
+      active={!item.external && isActive(item.href)}
+      collapsed={collapsed}
+      external={item.external}
+    />
+  )
 
   const renderGroupHeader = (label: string) =>
     !collapsed && (
@@ -176,39 +118,7 @@ export function Sidebar({ collapsed, onToggleCollapse }: SidebarProps) {
       >
         <nav className="flex-1 space-y-6 overflow-y-auto px-2 py-4">
           {showAdminNav ? (
-            <>
-              <div>
-                <div className="space-y-0.5">
-                  {renderNavLink({
-                    keyVal: 'back-to-site',
-                    href: '/',
-                    label: 'Back to site',
-                    icon: ArrowLeft,
-                    active: false,
-                  })}
-                </div>
-                <div className={cn('mt-2 border-t border-sidebar-border', collapsed ? 'mx-2' : 'mx-3')} />
-              </div>
-              {adminNavGroups.map(group => (
-                <div key={group.label}>
-                  {renderGroupHeader(group.label)}
-                  <div className="space-y-0.5">
-                    {group.items.map(item =>
-                      renderNavLink({
-                        keyVal: item.tab,
-                        href: adminTabHref(item.tab),
-                        label: item.label,
-                        icon: item.icon,
-                        active: isAdminTabActive(item.tab, pathname, tabParam),
-                        badge: item.badgeKey
-                          ? { count: counts[item.badgeKey], className: ADMIN_BADGE_CLASS[item.badgeKey] }
-                          : null,
-                      })
-                    )}
-                  </div>
-                </div>
-              ))}
-            </>
+            <AdminSidebarNav collapsed={collapsed} />
           ) : (
             <>
               {sidebarGroups.map(group => (
